+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 2.1.0 [unreleased]
+## 2.1.1 [unreleased]
+
+- Added `explain_timeout_sec` option
+- Fixed error with unparsable sequences
+
+## 2.1.0
 
 - Fixed issue with sequences in different schema than table
 - No longer throw errors for unreadable sequences